+++ changelog.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Multi diagrams page:
   * Focus movement on all diagrams.
+* Index page:
+  * Add a search bar to filter diagrams.
 
 ### Changed